refactor: reformat code
