+++ src/redux/actions/index.js
@@ -1,3 +1,3 @@
-export * from './async/authActions'
-export * from './sync/authActions'
-export * from './sync/shoppingBag.action'
+export * from './async/authenticationAsyncActions'
+export * from './sync/authenticationActions'
+export * from './sync/shoppingBagActions'
